requirements.txt
faker==37.1.0
kafka-python==2.1.2
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pytest==8.3.5
streamlit==1.44.1
pandas==2.2.3
